fix(Backend): remove errors schema

+++ backend/be-digital-booking/src/main/java/com/grupo9/db/service/LocationService.java
@@ -2,6 +2,7 @@ package com.grupo9.db.service;
 
 import com.grupo9.db.exceptions.BadRequestException;
 import com.grupo9.db.exceptions.ResourceNotFoundException;
+import com.grupo9.db.model.Category;
 import com.grupo9.db.model.Location;
 import com.grupo9.db.repository.ILocationRepository;
 import com.grupo9.db.util.ApiResponse;
@@ -23,13 +24,12 @@ public class LocationService {
         this.responsesBuilder = responsesBuilder;
     }
 
-    public ResponseEntity<ApiResponse> findAll(){
+    public ResponseEntity<ApiResponse<List<Location>, Object>> findAll(){
         List<Location> locations = iLocationRepository.findAll();
         return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Location List successfully",locations, null);
     }
 
-    public ResponseEntity<ApiResponse> findById(Long id) throws ResourceNotFoundException, BadRequestException {
-        if(id == null) throw new BadRequestException("ID missing");
+    public ResponseEntity<ApiResponse<Location, Object>> findById(Long id) throws ResourceNotFoundException {
         Optional<Location> location = iLocationRepository.findById(id);
         if(location.isEmpty()){
             throw new ResourceNotFoundException("Location with id " + id + " not found");
@@ -37,22 +37,33 @@ public class LocationService {
         return responsesBuilder.buildResponse(HttpStatus.OK.value(),"Get Location successfully", location.get(), null);
     }
 
-    public ResponseEntity<ApiResponse> save(Location location){
+    public ResponseEntity<ApiResponse<Location, Object>> save(Location location){
         Location response = iLocationRepository.save(location);
-        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Location created successfully", location, null);
+        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Location created successfully", response, null);
     }
 
-    public ResponseEntity<ApiResponse> update(Long id, Location location) throws ResourceNotFoundException, BadRequestException {
+    public ResponseEntity<ApiResponse<Location, Object>> update(Long id, Location location) throws ResourceNotFoundException, BadRequestException {
+        if(id == null) throw new BadRequestException("ID missing");
+
+        Boolean exists = iLocationRepository.existsById(id);
+        if(!exists){
+            throw new ResourceNotFoundException("Location with id " + id + " not found");
+        }
+
         location.setId(id);
-        this.findById(location.getId());
         Location response = iLocationRepository.save(location);
-        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Location updated successfully", location, null);
+        return responsesBuilder.buildResponse(HttpStatus.CREATED.value(),"Location updated successfully", response, null);
     }
 
     public ResponseEntity<ApiResponse> deleteById(Long id) throws ResourceNotFoundException, BadRequestException {
-        this.findById(id);
+        if(id == null) throw new BadRequestException("ID missing");
+        Boolean exists = iLocationRepository.existsById(id);
+        if(!exists){
+            throw new ResourceNotFoundException("Location with id " + id + " not found");
+        }
+
         iLocationRepository.deleteById(id);
-        return responsesBuilder.buildResponse(HttpStatus.NO_CONTENT.value(),"Location deleted successfully", "", null);
+        return responsesBuilder.buildResponse(HttpStatus.NO_CONTENT.value(),"Location deleted successfully", null, null);
     }
 
 }
